Trait bholdus_chainbridge_transfer::pallet::Config

pub trait Config: Config + Config {
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type Currency: MultiCurrency<Self::AccountId, CurrencyId = CurrencyId, Balance = Balance>;
    type BridgeOrigin: EnsureOrigin<Self::Origin, Success = Self::AccountId>;
    type AdminOrigin: EnsureOrigin<Self::Origin>;
    type NativeCurrencyId: Get<CurrencyId>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Associated Types

type Event: From<Event<Self>> + IsType<Self::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Currency: MultiCurrency<Self::AccountId, CurrencyId = CurrencyId, Balance = Balance>

Currency mechanism

type BridgeOrigin: EnsureOrigin<Self::Origin, Success = Self::AccountId>

Origin are allowed to transfer from the bridge

type AdminOrigin: EnsureOrigin<Self::Origin>

Origin are allowed to proceed admin operations

type NativeCurrencyId: Get<CurrencyId>

What is Native currency id

Loading content...